Add drag-and-drop functionality to order view
This commit introduces drag-and-drop capabilities to the `AddOrderVM` and `AddOrderView`. A new `DragDropBehavior` class is added to manage mouse events for item rearrangement in a `ListBox`. The `AddOrderVM` is updated with `DropCommand` and `DragData` properties to support this feature. The `ICloseable` interface is implemented to enable the `AddOrderView` to close itself. The XAML for `AddOrderView` is modified to incorporate the drag-and-drop behavior, improving user interaction.